+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -631,6 +631,54 @@ public class PostHog private constructor(
         capture(PostHogEventName.CREATE_ALIAS.event, properties = props)
     }
 
+    public override fun captureFeatureView(
+        flag: String,
+        flagVariant: String?,
+    ) {
+        if (!isEnabled()) {
+            return
+        }
+        val props = mutableMapOf<String, Any>()
+        props["feature_flag"] = flag
+
+        val variant = flagVariant ?: getFeatureFlag(flag, sendFeatureFlagEvent = false) ?: true
+        if (variant is String) {
+            props["feature_flag_variant"] = variant
+        }
+
+        val userProperties = mapOf("\$feature_view/$flag" to variant)
+
+        capture(
+            event = PostHogEventName.FEATURE_VIEW.event,
+            properties = props,
+            userProperties = userProperties,
+        )
+    }
+
+    public override fun captureFeatureInteraction(
+        flag: String,
+        flagVariant: String?,
+    ) {
+        if (!isEnabled()) {
+            return
+        }
+        val props = mutableMapOf<String, Any>()
+        props["feature_flag"] = flag
+
+        val variant = flagVariant ?: getFeatureFlag(flag, sendFeatureFlagEvent = false) ?: true
+        if (variant is String) {
+            props["feature_flag_variant"] = variant
+        }
+
+        val userProperties = mapOf("\$feature_interaction/$flag" to variant)
+
+        capture(
+            PostHogEventName.FEATURE_INTERACTION.event,
+            properties = props,
+            userProperties = userProperties,
+        )
+    }
+
     /**
      * Returns fresh default device and app properties for feature flag evaluation.
      */
@@ -1523,6 +1571,20 @@ public class PostHog private constructor(
 
         public override fun alias(alias: String) {
             shared.alias(alias)
+        }
+
+        public override fun captureFeatureView(
+            flag: String,
+            flagVariant: String?,
+        ) {
+            shared.captureFeatureView(flag, flagVariant)
+        }
+
+        public override fun captureFeatureInteraction(
+            flag: String,
+            flagVariant: String?,
+        ) {
+            shared.captureFeatureInteraction(flag, flagVariant)
         }
 
         public override fun isOptOut(): Boolean = shared.isOptOut()